Refactoring HTTP proxy so that it can be used for multiple API calls (and not just RAOIDC)
