+++ frontend/.env.example
@@ -40,9 +40,6 @@ AUTH_RAOIDC_BASE_URL=http://localhost:3000/auth
 # RAOIDC client ID -- the OAuth client id used for all RAOIDC calls
 # (required; use the example below)
 AUTH_RAOIDC_CLIENT_ID=CDCP
-# RAOIDC proxy URL -- used to proxy all outgoing RAOIDC calls
-# (optional; default: undefined)
-AUTH_RAOIDC_PROXY_URL=
 # RASCL logout URL -- used when no RAOIDC session is found during logout
 # (required; use the example below)
 AUTH_RASCL_LOGOUT_URL=http://localhost:3000/
@@ -65,6 +62,12 @@ HCAPTCHA_SITE_KEY=10000000-ffff-ffff-ffff-000000000001
 # hCaptcha verify URL -- used to verify hCaptcha tokens
 # (optional; default: https://api.hcaptcha.com/siteverify)
 HCAPTCHA_VERIFY_URL=https://api.hcaptcha.com/siteverify
+
+
+# HTTP proxy URL -- used to proxy specified outgoing HTTP calls
+# (optional; default: undefined)
+HTTP_PROXY_URL=
+
 
 # Interop API base url
 # (required; use the example below)
