+++ frontend/app/services/raoidc-service.server.ts
@@ -10,7 +10,7 @@
  *   - handleCallback() -- handles the OIDC callback request
  *   - handleLogout() -- handles an RAOIDC logout
  *
- * If an HTTP proxy is configured (via the AUTH_RAOIDC_PROXY_URL environment
+ * If an HTTP proxy is configured (via the HTTP_PROXY_URL environment
  * variable), the service will make all RAOIDC calls with a custom fetch()
  * function that uses that HTTP proxy. This is most commonly used during
  * development, when the nonprod RAOIDC instance is not accessible outside of
@@ -24,13 +24,12 @@ import { Session, redirect } from '@remix-run/node';
 
 import moize from 'moize';
 import { subtle } from 'node:crypto';
-import { ProxyAgent, fetch as undiciFetch } from 'undici';
-import { toNodeReadable } from 'web-streams-node';
 
 import { generateCryptoKey, generateJwkId } from '~/utils/crypto-utils.server';
 import { getEnv } from '~/utils/env.server';
+import { getFetchFn } from '~/utils/fetch-utils';
 import { getLogger } from '~/utils/logging.server';
-import type { ClientMetadata, FetchFunctionInit, IdToken, UserinfoToken } from '~/utils/raoidc-utils.server';
+import type { ClientMetadata, IdToken, UserinfoToken } from '~/utils/raoidc-utils.server';
 import { fetchAccessToken, fetchServerMetadata, fetchUserInfo, generateAuthorizationRequest, generateCodeChallenge, generateRandomState, validateSession } from '~/utils/raoidc-utils.server';
 import { expandTemplate } from '~/utils/string-utils';
 
@@ -45,8 +44,8 @@ export const getRaoidcService = moize.promise(createRaoidcService, { onCacheAdd:
  * Create and intialize an instance of the RAOID service.
  */
 async function createRaoidcService() {
-  const { AUTH_LOGOUT_REDIRECT_URL, AUTH_JWT_PRIVATE_KEY, AUTH_RAOIDC_BASE_URL, AUTH_RAOIDC_CLIENT_ID, AUTH_RAOIDC_PROXY_URL } = getEnv();
-  const fetchFn = getFetchFn(AUTH_RAOIDC_PROXY_URL);
+  const { AUTH_LOGOUT_REDIRECT_URL, AUTH_JWT_PRIVATE_KEY, AUTH_RAOIDC_BASE_URL, AUTH_RAOIDC_CLIENT_ID, HTTP_PROXY_URL } = getEnv();
+  const fetchFn = getFetchFn(HTTP_PROXY_URL);
   const { jwkSet, serverMetadata } = await fetchServerMetadata(AUTH_RAOIDC_BASE_URL, fetchFn);
 
   /**
@@ -140,24 +139,6 @@ async function createRaoidcService() {
     }
 
     log.debug('Authentication check passed');
-  }
-
-  /**
-   * Return a custom fetch() function if a proxy URL has been provided.
-   * If no proxy has been provided, simply return global.fetch().
-   */
-  function getFetchFn(proxyUrl?: string) {
-    if (proxyUrl) {
-      log.debug('A proxy has been configured: [%s]; using custom fetch for RAOIDC calls', proxyUrl);
-      return async (input: string | URL, init?: FetchFunctionInit) => {
-        const dispatcher = new ProxyAgent({ uri: proxyUrl, proxyTls: { timeout: 30000 } }); // TODO :: GjB :: make timeout configurable?
-        const response = await undiciFetch(input, { ...init, dispatcher });
-        return new Response(toNodeReadable(response.body));
-      };
-    }
-
-    log.debug('No proxy configured; using global fetch for RAOIDC calls');
-    return global.fetch;
   }
 
   return { generateSignoutRequest, generateSigninRequest, handleCallback, handleSessionValidation };
